+++ test/templates/react-srv-fn/CHANGELOG.md
@@ -4,6 +4,16 @@
 
 ### Patch Changes
 
+- Updated dependencies [ee9611f]
+- Updated dependencies [2dad740]
+  - vinxi@0.1.4
+  - @vinxi/server-functions@0.1.2
+  - @vinxi/react@0.1.1
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [583364f]
 - Updated dependencies [c45633c]
 - Updated dependencies [fd5740c]